refactor(nonce): separate amount to mod
