+++ rust/src/command/mod.rs
@@ -6,7 +6,7 @@ pub mod zkapp;
 use crate::{
     block::{precomputed::PrecomputedBlock, BlockHash},
     command::signed::{SignedCommand, SignedCommandWithKind},
-    ledger::{account::Nonce, amount::Amount, public_key::PublicKey},
+    ledger::{amount::Amount, nonce::Nonce, public_key::PublicKey},
     mina_blocks::v2::{
         self,
         staged_ledger_diff::{
