+++ rust/src/ledger/account/mod.rs
@@ -4,6 +4,7 @@ use super::{
         account::{AccountDiff, UpdateType},
         LedgerDiff,
     },
+    nonce::Nonce,
     token::TokenAddress,
     username::Username,
 };
@@ -14,68 +15,6 @@ use crate::{
     mina_blocks::v2::ZkappAccount,
 };
 use serde::{Deserialize, Serialize};
-use std::{
-    fmt::{self, Display},
-    ops::{Add, Sub},
-};
-
-#[derive(
-    PartialEq, Eq, Debug, Copy, Clone, Default, Serialize, Deserialize, PartialOrd, Ord, Hash,
-)]
-pub struct Nonce(pub u32);
-
-impl Add<u32> for Nonce {
-    type Output = Nonce;
-
-    fn add(self, other: u32) -> Nonce {
-        Self(self.0.saturating_add(other))
-    }
-}
-
-impl Sub<u32> for Nonce {
-    type Output = Nonce;
-
-    fn sub(self, other: u32) -> Nonce {
-        Self(self.0.saturating_sub(other))
-    }
-}
-
-impl Add<i32> for Nonce {
-    type Output = Nonce;
-
-    fn add(self, other: i32) -> Nonce {
-        let abs = other.unsigned_abs();
-        if other > 0 {
-            Self(self.0.saturating_add(abs))
-        } else {
-            Self(self.0.saturating_sub(abs))
-        }
-    }
-}
-
-impl From<u32> for Nonce {
-    fn from(value: u32) -> Self {
-        Self(value)
-    }
-}
-
-impl From<String> for Nonce {
-    fn from(value: String) -> Self {
-        Self(value.parse::<u32>().expect("nonce is u32"))
-    }
-}
-
-impl From<Nonce> for serde_json::value::Number {
-    fn from(value: Nonce) -> Self {
-        Self::from(value.0)
-    }
-}
-
-impl Display for Nonce {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
 
 #[derive(Default, Clone, PartialEq, Eq, Serialize, Deserialize)]
 pub struct Account {
