+++ rust/src/ledger/genesis.rs
@@ -1,6 +1,7 @@
 use super::{
-    account::{Account, Nonce, Permissions, ReceiptChainHash, Timing, TokenPermissions},
+    account::{Account, Permissions, ReceiptChainHash, Timing, TokenPermissions},
     amount::Amount,
+    nonce::Nonce,
     public_key::PublicKey,
     Ledger, TokenLedger,
 };
